Perform lifecycle aware data loading with loaders
For some unknown reason, the system calls the onLoadFinished method
before the onResume method when returning from the NoteActivity.
Following the app development instructions, this causes the app to
crash as it tries to access a closed database object.
Resolving this problem involves leaving the cursor open as described
in NoteRecyclerAdapter class.